+++ packages/components/src/common/dateUtils/index.js
@@ -3,3 +3,4 @@ export { moveToWithinRange } from './moveToWithinRange';
 export { isWithinRange } from './isWithinRange';
 export { getMonthNames } from './getMonthNames';
 export { isDateValid } from './isDateValid';
+export { MonthFormat } from './monthFormat';